daemon: move removeOldRouterState to infraIPAllocator
This commit refactors the function `removeOldRouterState` from `daemon.go`
to a method of the `infraIPAllocator`. This also includes moving the
tests.
